Extract TpcUtils class in product tests
